+++ src/models/Settings.php
@@ -8,6 +8,7 @@ class Settings extends Model
 {
     public $api_key;
     public $waitlist_id;
+    public $booking_length = 20;
 
     public function init()
     {
@@ -17,7 +18,7 @@ class Settings extends Model
     public function rules()
     {
         return [
-            [['api_key', 'waitlist_id'], 'required'],
+            [['api_key', 'waitlist_id', 'booking_length'], 'required'],
         ];
     }
 }