+++ src/models/Waitwhile.php
@@ -13,6 +13,7 @@ class Waitwhile extends Model
     public $guests = null;
     public $guestsWaiting = null;
     public $bookings = null;
+    public $bookingsFrom = null;
     public $resources = null;
     public $error = false;
     public $errors = [];
@@ -42,6 +43,7 @@ class Waitwhile extends Model
      * Fetch the waitwhile session
      *
      * @return mixed
+     * @throws \craft\errors\MissingComponentException
      */
     public function getSession()
     {
@@ -83,9 +85,7 @@ class Waitwhile extends Model
      */
     public function getWaitlist(): array
     {
-        $settings = $this->settings;
-
-        return \Craft::$app->cache->getOrSet("waitwhileWaitlist", function ($cache) use($settings) {
+        return \Craft::$app->cache->getOrSet("waitwhileWaitlist", function ($cache) {
             return $this->waitlist === null ? $this->makeRequest('waitlists/' . $this->settings->waitlist_id) : $this->waitlist;
         }, 300);
     }
@@ -95,9 +95,7 @@ class Waitwhile extends Model
      */
     public function getAllWaitlists(): array
     {
-        $settings = $this->settings;
-
-        return \Craft::$app->cache->getOrSet("waitwhileAllWaitlists", function ($cache) use($settings) {
+        return \Craft::$app->cache->getOrSet("waitwhileAllWaitlists", function ($cache) {
             return $this->waitlists === null ? $this->makeRequest('waitlists') : $this->waitlists;
         }, 300);
     }
@@ -107,9 +105,7 @@ class Waitwhile extends Model
      */
     public function getWaitlistStatus(): array
     {
-        $settings = $this->settings;
-
-        return \Craft::$app->cache->getOrSet("waitwhileWaitlistStatus", function ($cache) use($settings) {
+        return \Craft::$app->cache->getOrSet("waitwhileWaitlistStatus", function ($cache) {
             return $this->waitlistStatus === null ? $this->makeRequest('waitlists/' . $this->settings->waitlist_id . '/status') : $this->waitlistStatus;
         }, 300);
     }
@@ -119,9 +115,7 @@ class Waitwhile extends Model
      */
     public function getWaitingGuests(): array
     {
-        $settings = $this->settings;
-
-        return \Craft::$app->cache->getOrSet("waitwhileWaitingGuests", function ($cache) use($settings) {
+        return \Craft::$app->cache->getOrSet("waitwhileWaitingGuests", function ($cache) {
             return $this->guestsWaiting === null ? $this->makeRequest('waitlists/' . $this->settings->waitlist_id . '/waiting') : $this->guestsWaiting;
         }, 300);
     }
@@ -131,11 +125,19 @@ class Waitwhile extends Model
      */
     public function getBookings(): array
     {
-        $settings = $this->settings;
-
-        return \Craft::$app->cache->getOrSet("waitwhileBookings", function ($cache) use($settings) {
+        return \Craft::$app->cache->getOrSet("waitwhileBookings", function ($cache) {
             return $this->bookings === null ? $this->makeRequest('waitlists/' . $this->settings->waitlist_id . '/bookings') : $this->bookings;
         }, 300);
+    }
+
+    /**
+     * @param int $fromTime
+     * @return array
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     */
+    public function getBookingsFrom(int $fromTime = 0): array
+    {
+        return $this->bookingsFrom === null ? $this->makeRequest('waitlists/' . $this->settings->waitlist_id . '/bookings?fromTime=' . $fromTime) : $this->bookings;
     }
 
     /**
@@ -168,5 +170,92 @@ class Waitwhile extends Model
         return \Craft::$app->cache->getOrSet("waitwhileResources", function ($cache) use($settings) {
             return $this->resources === null ? $this->makeRequest('resources') : $this->resources;
         }, 300);
+    }
+
+    /**
+     * @param string $date
+     * @return array
+     * @throws \GuzzleHttp\Exception\GuzzleException
+     * @throws \yii\base\ExitException
+     */
+    public function getBookingTimesForDay(string $date): array
+    {
+        // start of $date
+        $startOfDate = (new \DateTime($date))->setTime(0,0);
+        $startOfDateUnixMs = $startOfDate->getTimestamp() * 1000;
+        $currentTimeUnixMs = (new \DateTime())->getTimestamp() * 1000;
+
+        // bookings from start of today
+        $bookings = $this->getBookingsFrom($startOfDateUnixMs);
+
+        $waitlistHours = $this->getWaitlist()['waitlistHours'];
+        $currentDayAbbreviation = strtolower($startOfDate->format('D'));
+        $waitlistHoursToday = $waitlistHours[$currentDayAbbreviation]['periods'];
+        $isOpen = $waitlistHours[$currentDayAbbreviation]['isOpen'];
+
+        $bookingLength = $this->settings['booking_length'];
+
+        $times = [];
+
+        $bookingLengthUnixMs = $bookingLength * 60 * 1000;
+        $currentTimeUnixMsOnlyToday = $currentTimeUnixMs - $startOfDateUnixMs;
+
+        foreach($waitlistHoursToday as $period){
+            // we need to foreach each period to create available times within them based on $bookingLengthUnixMs
+
+            $start = $period['from'];
+            $end = $period['to'];
+
+            while($start < $end){
+                $realTimeUnixMs = $startOfDateUnixMs + $start;
+                $nextStart = $start + $bookingLengthUnixMs;
+                $nextStartRealTime = $startOfDateUnixMs + $nextStart;
+
+                // $currentTimeUnixMsOnlyToday must be less than $start
+                if($currentTimeUnixMsOnlyToday < $start) {
+                    // we need to check if available based on $bookings array
+                    $available = true;
+
+                    foreach($bookings as $booking) {
+                        // check if $booking['time'] is not the same as or within $realTimeUnixMs and $nextStart
+                        if($booking['time'] >= $realTimeUnixMs && $booking['time'] < $nextStartRealTime) $available = false;
+                    }
+
+                    // logic to add to $times array
+                    $times[] = [
+                        'start' => self::unix_ms_to_human($start),
+                        'start_unix_ms' => $start,
+                        'duration' => $bookingLengthUnixMs,
+                        'available' => $available
+                    ];
+                }
+
+                // increment $start with the length of $bookingLengthUnixMs
+                $start = $nextStart;
+            }
+        }
+
+        return [
+            'isOpen' => $isOpen,
+            'times' => $times,
+        ];
+    }
+
+    /**
+     * @param $value
+     * @return false|string
+     */
+    public static function unix_ms_to_human($value)
+    {
+        return gmdate("H:i", $value / 1000);
+    }
+
+    /**
+     * @param $value
+     * @return false|string
+     */
+    public static function unix_ms_to_minutes($value)
+    {
+        return round($value / 1000 / 60);
     }
 }