+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Everyday Waitwhile
 
+## 1.1.0 - 2018-09-10
+- Added a `Waitwhile@getBookingTimesForDay('YYYY-MM-DD')` method that returns an array with the available times on the given day
+- Added a `Booking::formatBookins` method, only used internally
+- Added a booking length setting (in minutes)
+
 ## 1.0.6 - 2018-09-07
 - Added support for a hidden input field called `country_code` in submission for action `everyday-waitwhile/queue`
 (input is without +, only the actual country code)